MAT-6038 test case import validation and refactor
